Add a Drop implementation to the pnet PingTransport
Without this change, the receiver threads should clean themselves up
eventually; if matching ICMP echo are received and attempted to be sent.
This attempts to have the threads close as soon as the transport is
dropped by sending a matching packet to the loopback address.

It does not change correctness, just efficiency.