Let PingTransport fully own the TransportSender
This makes it easier to reason about ownership (at least in that part of
the stack) at compile time, and avoids lock operations for every packet.

The `Pinger` changes to maintaining an Arc<Mutex<impl Transport>>, so
that one transport could have references from multiple Pingers.

The downside is that in the case that we do send from multiple threads,
we take one coarse grained lock for the duration of the send, so calls
to `run_pings` might end up taking longer to start, though total time
should be roughly equivalent.

Since the TransportSender needs a mutable reference to itself to do the
send, we need to plumb through some `&mut` changes.