Use a Typestate pattern to signal a running Pinger
This allows the Pinger to take ownership of the Receiver for the
transport and the Transport itself, which are the last things that,
semantically, it should own.

In some ways this makes the API harder to understand, but it also more
clearly signals which pieces of the code own which data structures, and
what the lifetimes are, making it harder to make mistakes.